+++ engine/Macro.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#define SCREEN_WIDTH 1920
+#define SCREEN_HEIGHT 1080
+
+#define SFML_LOGO "res/sfml-logo-big-word.png"
+#define LOGO_SFX "res/LogoSFX.ogg"
+
+#define HERO_ANIMATION_SPRITESHEET_R "res/Adventurer/adventurer-Sheet.png"
+#define HERO_ANIMATION_SPRITESHEET_L "res/Adventurer/adventurer-Sheet-L.png"
+
+#define LOG_FILE_PATH "logs/gamelog.log"
