+++ game/SplashState.cpp
@@ -1,65 +1,69 @@
 #include "SplashState.h"
 #include "DemoState.h"
 
-#include "controls/Input.h"
-#include "ui/UI_Button.h"
+#include "../controls/Input.h"
+#include "../ui/UI_Button.h"
 
 //namespace LUCY
 //{
-	/*SplashState::SplashState(GameDataRef data): m_data(data) , m_hero(data)
-	{
-	}
-
-	void SplashState::VInit()
-	{
-		Input::get().registerKey("Hello", sf::Keyboard::Space);
-
-		button.setPosition(sf::Vector2f(10, 10));
-		button.setColor(sf::Color::White, sf::Color::Blue, sf::Color::Yellow, sf::Color::Red, sf::Color::Green);
-		button.setBounds(sf::Vector2f(100, 50));
-
-		button1.setPosition(sf::Vector2f(10, 100));
-		button1.setColor(sf::Color::Yellow, sf::Color::Blue, sf::Color::Yellow, sf::Color::White, sf::Color::Red);
-		button1.setBounds(sf::Vector2f(100, 50));
-
-		std::cout << button.getID() << std::endl;
-		std::cout << button1.getID() << std::endl;
-
-		camera.set(sf::Vector2f(640 / 2, 480 / 2), sf::Vector2f(640, 480));
-		camera.setRenderTarget(&m_data->window);
-		camera.translateCameraToPosition(sf::Vector2f (320, 255), 0.1f);
-	}
-	void SplashState::VDraw(float dt)
-	{
-		m_data->window.clear();
-		button.draw(m_data->window);
-		button1.draw(m_data->window);
-
-		m_data->window.display();
-	}
-	void SplashState::VHandleInput()*/
+//	SplashState::SplashState(GameDataRef data) : m_data(data), m_hero(data)
+//	{
+//	}
+//
+//	void SplashState::VInit()
+//	{
+//		Input::get().registerKey("Hello", sf::Keyboard::Space);
+//
+//		button.setPosition(sf::Vector2f(10, 10));
+//		button.setColor(sf::Color::White, sf::Color::Blue, sf::Color::Yellow, sf::Color::Red, sf::Color::Green);
+//		button.setBounds(sf::Vector2f(100, 50));
+//
+//		button1.setPosition(sf::Vector2f(10, 100));
+//		button1.setColor(sf::Color::Yellow, sf::Color::Blue, sf::Color::Yellow, sf::Color::White, sf::Color::Red);
+//		button1.setBounds(sf::Vector2f(100, 50));
+//
+//		std::cout << button.getID() << std::endl;
+//		std::cout << button1.getID() << std::endl;
+//
+//		camera.set(sf::Vector2f(640 / 2, 480 / 2), sf::Vector2f(640, 480));
+//		camera.setRenderTarget(&m_data->window);
+//		camera.translateCameraToPosition(sf::Vector2f(320, 255), 0.1f);
+//	}
+//
+//	void SplashState::VDraw(float dt)
+//	{
+//		m_data->window.clear();
+//		button.draw(m_data->window);
+//		button1.draw(m_data->window);
+//
+//		m_data->window.display();
+//	}
+//
+//	void SplashState::VHandleInput()
+//	{
+//
+//	}
+//}
   
-LUCY::SplashState::SplashState(GameDataRef data): m_data(data)
+LUCY::SplashState::SplashState(GameDataRef data): m_data(data), m_hero(data)
 {
 }
 
 void LUCY::SplashState::VInit()
 {
-	logo_sfx.setBuffer(m_data->assets.GetSoundBuffer("LOGO_SFX"));
-	logo_sfx.play();
+	/*logo_sfx.setBuffer(m_data->assets.GetSoundBuffer("LOGO_SFX"));
+	logo_sfx.play();*/
 	m_data->assets.LoadTexture("SFML_Logo", SFML_LOGO);
 
 	sf_logo.setTexture(m_data->assets.GetTexture("SFML_Logo"));
 	sf_logo.setScale(0.5, 0.5);
 	sf_logo.setPosition(sf::Vector2f(SCREEN_WIDTH / 2, SCREEN_HEIGHT / 2));
 
-	//sf_logo.setOrigin(sf::Vector2f(sf_logo.getGlobalBounds().width, sf_logo.getGlobalBounds().height));	//final
+	sf_logo.setOrigin(sf::Vector2f(sf_logo.getGlobalBounds().width, sf_logo.getGlobalBounds().height));
 	stop_point.x = SCREEN_WIDTH/2 - sf_logo.getGlobalBounds().width / 3;
-	sf_logo.setOrigin(sf::Vector2f(sf_logo.getGlobalBounds().width / 3, sf_logo.getGlobalBounds().height)); //start
-	/*sf_logo.setPosition(sf::Vector2f(SCREEN_WIDTH / 2 - sf_logo.getGlobalBounds().width/2,
-		SCREEN_HEIGHT / 2 - sf_logo.getGlobalBounds().height/2));*/
-	
-	
+	sf_logo.setOrigin(sf::Vector2f(sf_logo.getGlobalBounds().width / 3, sf_logo.getGlobalBounds().height));
+	sf_logo.setPosition(sf::Vector2f(SCREEN_WIDTH / 2 - sf_logo.getGlobalBounds().width/2,
+		SCREEN_HEIGHT / 2 - sf_logo.getGlobalBounds().height/2));
 
 	rect_mask.setSize(sf::Vector2f(500.0f, 128.0f));
 	rect_mask.setPosition(sf::Vector2f(sf_logo.getPosition().x + 75.f, sf_logo.getPosition().y - 50.0f));
@@ -105,23 +109,22 @@ void LUCY::SplashState::VHandleInput()
 			}
 		}
 	}
-	void SplashState::VUpdate(float dt)
-	{
-		button.update(m_data->window);
-		button1.update(m_data->window);
-
-		camera.update();
-
-		//std::cout << m_data->window.getView().getCenter().x << " " << m_data->window.getView().getCenter().y << std::endl;
-			this->VExit();
-			this->m_data->window.close();
-		}
-		if (sf::Keyboard::isKeyPressed(sf::Keyboard::Escape))
-		{
-			this->VExit();
-		}
-	}
 }
+//	void SplashState::VUpdate(float dt)
+//	{
+//		button.update(m_data->window);
+//		button1.update(m_data->window);
+//
+//		camera.update();
+//
+//		//std::cout << m_data->window.getView().getCenter().x << " " << m_data->window.getView().getCenter().y << std::endl;
+//			this->VExit();
+//			this->m_data->window.close();
+//		if (sf::Keyboard::isKeyPressed(sf::Keyboard::Escape))
+//		{
+//			this->VExit();
+//		}
+//}
 
 void LUCY::SplashState::VUpdate(float dt)
 {
